record src version info with the built artifact

+++ builder.sh
@@ -3,9 +3,11 @@ set -e
 
 fatal(){
   echo "$1"
-  echo "Usage: $0 [debug|normal] [newlib|linux] [compiler|tools|clean|fesvr|spike|dpi|pk]"
+  echo "Usage: $0 [debug|normal] [newlib|linux] [compiler|tools|clean|fesvr|spike|dpi|pk|version_only]"
   exit 1
 }
+
+CR=$'\n'
 
 if [[ $# != 3 ]]; then
   fatal "please provides 3 argument"
@@ -22,6 +24,7 @@ TARGET_FESVR="no"
 TARGET_SPIKE="no"
 TARGET_DPI="no"
 TARGET_PK="no"
+TARGET_VERSION_ONLY="no"
 
 
 for tgt in "${BUILD_TARGETS[@]}"; do
@@ -45,6 +48,8 @@ for tgt in "${BUILD_TARGETS[@]}"; do
     TARGET_DPI="yes"
   elif [[ $tgt == "pk" && x$TARGET_TOOLS == "xno" ]]; then
     TARGET_PK="yes"
+  elif [[ $tgt == "version_only" && x$TARGET_VERSION_ONLY == "xno" ]]; then
+    TARGET_VERSION_ONLY="yes"
   else
     fatal "Error: bad target: $tgt"
   fi
@@ -78,6 +83,9 @@ else
   fatal "Error: bad toolchain: $BUILD_TOOLCHAIN"
 fi
 
+source ./build.common
+source ./version.common
+
 echo "******************************************"
 echo "                Task brief                "
 echo "******************************************"
@@ -96,9 +104,7 @@ echo "   DPI:             *    $TARGET_DPI      "
 echo "   PRORY KERNEL:    *    $TARGET_PK       "
 echo "******************************************"
 
-source ./build.common
-
-if [[ x$TARGET_CLEAN == "xyes" ]]; then
+if [[ x$TARGET_CLEAN == "xyes" && x$TARGET_VERSION_ONLY == "xno" ]]; then
   echo "Cleaning RISC-V Compiler and Tools installation ($RISCV_INSTALL)."
 
   rm -fr $RISCV_INSTALL
@@ -109,29 +115,47 @@ fi
 if [[ x$TARGET_COMPILER == "xyes" ]]; then
   echo "Starting RISC-V Compiler build process"
   if [[ $BUILD_TOOLCHAIN == "newlib" ]]; then
-    CXXFLAGS_FOR_TARGET_EXTRA="-g" CFLAGS_FOR_TARGET_EXTRA="-g" build_gcc_newlib riscv-gnu-toolchain --prefix=$RISCV_INSTALL --with-arch=rv64imfd  --with-abi=lp64d
+    if [[ x$TARGET_VERSION_ONLY == "xno" ]]; then
+      CXXFLAGS_FOR_TARGET_EXTRA="-g" CFLAGS_FOR_TARGET_EXTRA="-g" build_gcc_newlib riscv-gnu-toolchain --prefix=$RISCV_INSTALL --with-arch=rv64imfd  --with-abi=lp64d
+    fi
+    log_newlib_toolchain_version_to "$RISCV_INSTALL"
   else
-    CXXFLAGS_FOR_TARGET_EXTRA="-g" CFLAGS_FOR_TARGET_EXTRA="-g" build_gcc_linux  riscv-gnu-toolchain --prefix=$RISCV_INSTALL --with-arch=rv64imafd --with-abi=lp64d
+    if [[ x$TARGET_VERSION_ONLY == "xno" ]]; then
+      CXXFLAGS_FOR_TARGET_EXTRA="-g" CFLAGS_FOR_TARGET_EXTRA="-g" build_gcc_linux  riscv-gnu-toolchain --prefix=$RISCV_INSTALL --with-arch=rv64imafd --with-abi=lp64d
+    fi
+    log_linux_toolchain_version_to "$RISCV_INSTALL"
   fi
 
   echo -e "\\nRISC-V Compiler installation completed!"
 fi
 
 if [[ x$TARGET_FESVR == "xyes" ]]; then
-  build_project riscv-fesvr --prefix=$RISCV_INSTALL
+  if [[ x$TARGET_VERSION_ONLY == "xno" ]]; then
+    build_project riscv-fesvr --prefix=$RISCV_INSTALL
+  fi
+  log_fesvr_version_to "$RISCV_INSTALL"
 fi
 
 if [[ x$TARGET_SPIKE == "xyes" ]]; then
-  build_project riscv-isa-sim --prefix=$RISCV_INSTALL --with-fesvr=$RISCV_INSTALL --enable-simpoint $spike_extra_flag
+  if [[ x$TARGET_VERSION_ONLY == "xno" ]]; then
+    build_project riscv-isa-sim --prefix=$RISCV_INSTALL --with-fesvr=$RISCV_INSTALL --enable-simpoint $spike_extra_flag
+  fi
+  log_spike_version_to "$RISCV_INSTALL" "${CR}For $BUILD_TYPE use"
 fi
 
 if [[ x$TARGET_DPI == "xyes" ]]; then
-  build_project riscv-dpi --prefix=$RISCV_INSTALL --with-fesvr=$RISCV_INSTALL --enable-checker
+  if [[ x$TARGET_VERSION_ONLY == "xno" ]]; then
+    build_project riscv-dpi --prefix=$RISCV_INSTALL --with-fesvr=$RISCV_INSTALL --enable-checker
+  fi
+  log_dpi_version_to "$RISCV_INSTALL"
 fi
 
 if [[ x$TARGET_PK == "xyes" ]]; then
   PATH="$NEWLIB_GCC_PATH/bin:$PATH" check_newlib_gcc
-  PATH="$NEWLIB_GCC_PATH/bin:$PATH" CC=riscv64-unknown-elf-gcc CFLAGS="-g -D__riscv64 -march=rv64imfd -mabi=lp64d"  ASFLAGS="-march=rv64imfd -mabi=lp64d" build_project riscv-pk --prefix=$RISCV_INSTALL/riscv64-unknown-elf --host=riscv --disable-atomics
+  if [[ x$TARGET_VERSION_ONLY == "xno" ]]; then
+    PATH="$NEWLIB_GCC_PATH/bin:$PATH" CC=riscv64-unknown-elf-gcc CFLAGS="-g -D__riscv64 -march=rv64imfd -mabi=lp64d"  ASFLAGS="-march=rv64imfd -mabi=lp64d" build_project riscv-pk --prefix=$RISCV_INSTALL/riscv64-unknown-elf --host=riscv --disable-atomics
+  fi
+  log_pk_version_to "$RISCV_INSTALL" "${CR}Build by riscv64-unknown-elf-toolchain:${CR}$(cat $NEWLIB_GCC_PATH/version/riscv64-unknown-elf-toolchain)"
 fi
 
 echo -e "\\nCompleted!"
